12/9/2023 8:55am
Everything part of the basic website foundation is all sorted now. All the bootstrap is functional, including the side offCanvas menu, the filters, etc.  We can work on the colors and aesthetics later, the web app's basic function is a higher priority.

The front-end things yet to be done: (off the top of my head)
- The table for the index in side menu
- Haven't added the feature for the side menu so that if the window is sm or smaller then the offcanvas will take up the whole screen instead of 75%
- The exact format for the stuff in location info and comments
- A pop up login menu when the login button is clicked
- An admin panel